refactor : 사다리 게임 결과 dto 인스턴스 필드 개수 줄이기
