+++ src/main/java/laddergame/dto/LadderGameResult.java
@@ -1,41 +1,36 @@
 package laddergame.dto;
 
-import laddergame.domain.LadderGame;
+import laddergame.domain.Ladder;
 import laddergame.domain.LadderGameReward;
 import laddergame.domain.ParticipantName;
-import laddergame.domain.ParticipantNames;
 
-import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
-import java.util.stream.IntStream;
+import java.util.stream.Collectors;
 
 public class LadderGameResult {
 
-    private final LadderGameDto ladderGame;
+    private final LadderDto ladder;
     private final Map<String, String> rewardByName;
 
-    public LadderGameResult(LadderGameDto ladderGame, Map<String, String> rewardByName) {
-        this.ladderGame = ladderGame;
+    public LadderGameResult(LadderDto ladder, Map<String, String> rewardByName) {
+        this.ladder = ladder;
         this.rewardByName = rewardByName;
     }
 
-    public static LadderGameResult from(LadderGame ladderGame) {
-        LadderGameDto ladderGameDto = LadderGameDto.from(ladderGame);
-        Map<ParticipantName, LadderGameReward> rewardByName = ladderGame.calculateRewardByParticipantName();
-        Map<String, String> rewardByNameDto = new HashMap<>();
-        ParticipantNames participantNames = ladderGame.getParticipantNames();
-        IntStream.range(0, participantNames.size())
-                .mapToObj(participantNames::get)
-                .forEach(participantName -> rewardByNameDto.put(
-                        participantName.toString(),
-                        rewardByName.get(participantName).toString())
-                );
-        return new LadderGameResult(ladderGameDto, rewardByNameDto);
+    public static LadderGameResult of(Ladder ladder, Map<ParticipantName, LadderGameReward> rewardByName) {
+        LadderDto ladderDto = LadderDto.from(ladder);
+        Map<String, String> rewardByNameDto = rewardByName.keySet()
+                .stream()
+                .collect(Collectors.toMap(
+                        ParticipantName::toString,
+                        participantName -> rewardByName.get(participantName).toString()
+                ));
+        return new LadderGameResult(ladderDto, rewardByNameDto);
     }
 
-    public LadderGameDto getLadderGame() {
-        return ladderGame;
+    public LadderDto getLadder() {
+        return ladder;
     }
 
     public Map<String, String> getRewardByName() {
@@ -47,18 +42,18 @@ public class LadderGameResult {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         LadderGameResult that = (LadderGameResult) o;
-        return Objects.equals(ladderGame, that.ladderGame) && Objects.equals(rewardByName, that.rewardByName);
+        return Objects.equals(ladder, that.ladder) && Objects.equals(rewardByName, that.rewardByName);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(ladderGame, rewardByName);
+        return Objects.hash(ladder, rewardByName);
     }
 
     @Override
     public String toString() {
         return "LadderGameResult{" +
-                "ladderGame=" + ladderGame +
+                "ladder=" + ladder +
                 ", rewardByName=" + rewardByName +
                 '}';
     }
