+++ src/main/java/laddergame/view/OutputView.java
@@ -3,7 +3,7 @@ package laddergame.view;
 import laddergame.domain.AbstractLadderGameValue;
 import laddergame.domain.ParticipantName;
 import laddergame.dto.LadderDto;
-import laddergame.dto.LadderGameDto;
+import laddergame.dto.LadderGameRunRequest;
 import laddergame.dto.LadderLineDto;
 
 import java.util.List;
@@ -20,10 +20,10 @@ public class OutputView {
     private OutputView() {
     }
 
-    public static void printLadderGame(LadderGameDto ladderGame) {
-        printParticipantNames(ladderGame.getParticipantNames());
-        printLadder(ladderGame.getLadder());
-        printRewards(ladderGame.getRewards());
+    public static void printLadderGame(LadderGameRunRequest request, LadderDto ladder) {
+        printParticipantNames(request.getParticipantNames());
+        printLadder(ladder);
+        printRewards(request.getRewards());
     }
 
     private static void printParticipantNames(List<String> participantNames) {
